4. cart/smartM_ai_server/requirements.txt
altgraph==0.17.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
comtypes==1.1.11
cycler==0.11.0
distlib==0.3.4
docutils==0.18.1
easydict==1.9
filelock==3.6.0
firebase-admin==5.2.0
Flask==2.1.1
fonttools==4.33.3
future==0.18.2
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
googletrans==3.0.0
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
msgpack==1.0.4
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pefile==2021.9.3
Pillow==9.1.0
pipenv==2022.4.21
platformdirs==2.5.2
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.3.0
pynput==1.7.6
pyparsing==3.0.8
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyserial==3.5
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
pyupbit==0.2.24
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
schedule==1.1.0
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
textblob==0.17.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.64.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.14.1
virtualenv-clone==0.5.7
websockets==9.1
Werkzeug==2.1.1
wget==3.2
zipp==3.8.0
